refactor: transform airline module
