+++ src/modules/airline/index.ts
@@ -1,1 +1,12 @@
+/**
+ * IATA stands for [International Air Transport Association](https://iata.org).
+ * It's the trade association for the world's airlines and it is
+ * responsible for setting standards relating to many aspects of airline
+ * operations.
+ */
+export { Aircraft } from './aircraft-type';
+export type { AircraftType } from './aircraft-type';
+export type { Airline } from './airline';
+export type { Airplane } from './airplane';
+export type { Airport } from './airport';
 export * from './module';
